+++ modules/quadbin/snowflake/sql/QUADBIN_RESOLUTION.sql
@@ -3,7 +3,7 @@
 ----------------------------
 
 CREATE OR REPLACE FUNCTION QUADBIN_RESOLUTION
-(quadbin INT)
+(quadbin BIGINT)
 RETURNS INT
 AS $$
     SELECT BITAND(BITSHIFTRIGHT(quadbin, 52), 31)
