+++ modules/quadbin/snowflake/sql/QUADBIN_TOPARENT.sql
@@ -3,19 +3,19 @@
 ----------------------------
 
 CREATE OR REPLACE FUNCTION QUADBIN_TOPARENT
-(quadbin INT, resolution INT)
-RETURNS INT
+(quadbin BIGINT, resolution INT)
+RETURNS BIGINT
 AS $$
     SELECT bitor(
                 bitor(
                     bitand(
-                        quadbin, 
+                        quadbin,
                         bitnot(
                             bitshiftleft(31, 52)
-                        )                        
+                        )
                     ),
                     bitshiftleft(resolution, 52)
-                ), 
+                ),
                 bitshiftright(
                     4503599627370495,
                     resolution * 2
